+++ src/hexen/r_things.c
@@ -52,8 +52,8 @@ lighttable_t **spritelights;
 extern fixed_t          fractionaltic;
 
 // constant arrays used for psprite clipping and initializing clipping
-short negonearray[MAXWIDTH];
-short screenheightarray[MAXWIDTH];
+int negonearray[MAXWIDTH];  // [crispy] 32-bit integer math
+int screenheightarray[MAXWIDTH];  // [crispy] 32-bit integer math
 
 boolean LevelUseFullBright;
 /*
@@ -319,8 +319,8 @@ vissprite_t *R_NewVisSprite(void)
 ================
 */
 
-short *mfloorclip;
-short *mceilingclip;
+int *mfloorclip;  // [crispy] 32-bit integer math
+int *mceilingclip;  // [crispy] 32-bit integer math
 fixed_t spryscale;
 fixed_t sprtopscreen;
 fixed_t sprbotscreen;
@@ -994,7 +994,7 @@ void R_SortVisSprites(void)
 void R_DrawSprite(vissprite_t * spr)
 {
     drawseg_t *ds;
-    short clipbot[MAXWIDTH], cliptop[MAXWIDTH];
+    int clipbot[MAXWIDTH], cliptop[MAXWIDTH];  // [crispy] 32-bit integer math
     int x, r1, r2;
     fixed_t scale, lowscale;
     int silhouette;
